example codes bugfixes (/bin) improved control words (mainly for Quickstop) actual position homing method (35) fixed rounding error added motor data parameters (max_speed and thermal time constant) fixed position_profile_set_velocity added profile max velocity, max acceleration and quickstop settings

+++ src/lib/gear.c
@@ -40,7 +40,8 @@ float epos_gear_to_angle(const epos_gear_t* gear, int position) {
 }
 
 int epos_gear_from_angle(const epos_gear_t* gear, float angle) {
-  return clip(roundf(angle*4.0*gear->sensor->num_pulses*gear->transmission/(2.0*M_PI)),
+  return clip(round(angle*4.0*gear->sensor->num_pulses*gear->transmission/
+                    (2.0*M_PI)),
     INT_MIN, INT_MAX);
 }
 
@@ -50,7 +51,7 @@ float epos_gear_to_angular_velocity(const epos_gear_t* gear, int velocity) {
 
 int epos_gear_from_angular_velocity(const epos_gear_t* gear, float
     angular_vel) {
-  return clip(roundf(angular_vel*60.0*gear->transmission/(2.0*M_PI)),
+  return clip(round(angular_vel*60.0*gear->transmission/(2.0*M_PI)),
     INT_MIN, INT_MAX);
 }
 
@@ -61,6 +62,6 @@ float epos_gear_to_angular_acceleration(const epos_gear_t* gear, int
 
 int epos_gear_from_angular_acceleration(const epos_gear_t* gear, float
     angular_acc) {
-  return clip(roundf(angular_acc*60.0*abs(gear->transmission)/(2.0*M_PI)),
+  return clip(round(angular_acc*60.0*abs(gear->transmission)/(2.0*M_PI)),
     INT_MIN, INT_MAX);
 }
